fix: authenticate user before CRUD the review

+++ api/service/review.service.go
@@ -11,28 +11,37 @@ var (
 )
 
 type ReviewService interface {
-	CreateReview(ctx context.Context, rating int, comment string, artToyID int64) (*domain.Review, error)
+	CreateReview(ctx context.Context, rating int, comment string, artToyID int64, ownerID int64) (*domain.Review, error)
 	GetReview(ctx context.Context, artToyID int64) (*domain.Review, error)
-	UpdateReview(ctx context.Context, artToyID int64, updateBody map[string]any) (*domain.Review, error)
-	DeleteReview(ctx context.Context, artToyID int64) error
+	UpdateReview(ctx context.Context, artToyID int64, updateBody map[string]any, ownerID int64) (*domain.Review, error)
+	DeleteReview(ctx context.Context, artToyID int64, ownerID int64) error
 }
 
 type reviewServiceImpl struct {
+	orderRepo  domain.OrderRepository
 	reviewRepo domain.ReviewRepository
 }
 
-func NewReviewService(reviewRepo domain.ReviewRepository) ReviewService {
+func NewReviewService(orderRepo domain.OrderRepository, reviewRepo domain.ReviewRepository) ReviewService {
 	return &reviewServiceImpl{
+		orderRepo,
 		reviewRepo,
 	}
 }
 
 var _ ReviewService = &reviewServiceImpl{}
 
-func (s *reviewServiceImpl) CreateReview(ctx context.Context, rating int, comment string, artToyID int64) (*domain.Review, error) {
-	review := domain.NewReview(rating, comment, artToyID)
-	err := s.reviewRepo.CreateReview(ctx, review)
+func (s *reviewServiceImpl) CreateReview(ctx context.Context, rating int, comment string, artToyID int64, ownerID int64) (*domain.Review, error) {
+	buyerID, err := s.reviewRepo.FindReviewerByArtToyID(ctx, artToyID)
 	if err != nil {
+		return nil, err
+	}
+	if *buyerID != ownerID {
+		return nil, ErrUnauthorized
+	}
+
+	review := domain.NewReview(rating, comment, artToyID)
+	if err := s.reviewRepo.CreateReview(ctx, review); err != nil {
 		return nil, err
 	}
 	return review, nil
@@ -46,9 +55,16 @@ func (s *reviewServiceImpl) GetReview(ctx context.Context, artToyID int64) (*dom
 	return review, nil
 }
 
-func (s *reviewServiceImpl) UpdateReview(ctx context.Context, artToyID int64, updateBody map[string]any) (*domain.Review, error) {
-	err := s.reviewRepo.UpdateReview(ctx, artToyID, updateBody)
+func (s *reviewServiceImpl) UpdateReview(ctx context.Context, artToyID int64, updateBody map[string]any, ownerID int64) (*domain.Review, error) {
+	buyerID, err := s.reviewRepo.FindReviewerByArtToyID(ctx, artToyID)
 	if err != nil {
+		return nil, err
+	}
+	if *buyerID != ownerID {
+		return nil, ErrUnauthorized
+	}
+
+	if err := s.reviewRepo.UpdateReview(ctx, artToyID, updateBody); err != nil {
 		return nil, err
 	}
 	updatedReview, err := s.reviewRepo.FindReviewByArtToyID(ctx, artToyID)
@@ -58,6 +74,13 @@ func (s *reviewServiceImpl) UpdateReview(ctx context.Context, artToyID int64, up
 	return updatedReview, nil
 }
 
-func (s *reviewServiceImpl) DeleteReview(ctx context.Context, artToyID int64) error {
+func (s *reviewServiceImpl) DeleteReview(ctx context.Context, artToyID int64, ownerID int64) error {
+	buyerID, err := s.reviewRepo.FindReviewerByArtToyID(ctx, artToyID)
+	if err != nil {
+		return err
+	}
+	if *buyerID != ownerID {
+		return ErrUnauthorized
+	}
 	return s.reviewRepo.DeleteReview(ctx, artToyID)
 }
